+++ changelog.md
@@ -5,7 +5,7 @@
 
 ## 2.2
 
-* [Changed] Some data are stored to `sample-data` instead of `sample_data`
+* [Changed] Some data are stored to `sample_data` instead of `sample-data`
 * [Changed] Internal representation of dataset dir is now just `root` 
 (instead of`root/medical/orig`)